ADvisor/requirements.txt
flask
flask-cors
pillow
moondream
requests
